✨ LXDE configuration!
Due to the fact that my little ubuntu machine is a slow ass
piece of SHIT. I've opted to the lighter LXDE display manager.

Note: There's a couple of funky custom scripts in there.

Mostly lxlock (should come with lxde) and movescreen.
Movescreen is python script that moves the screen (duh)

I __should__ probably make this a .patch file. But yolo 😎